Refactor of QueueResponseProcess to remove superfluous strategy layer.
