+++ org.eclipse.scanning.event.queues/src/org/eclipse/scanning/event/queues/remote/GetServerStringResponse.java
@@ -1,6 +1,7 @@
 package org.eclipse.scanning.event.queues.remote;
 
 import org.eclipse.scanning.api.event.EventException;
+import org.eclipse.scanning.api.event.core.IPublisher;
 import org.eclipse.scanning.api.event.queues.IQueueControllerService;
 import org.eclipse.scanning.api.event.queues.IQueueService;
 import org.eclipse.scanning.api.event.queues.remote.QueueRequest;
@@ -13,16 +14,17 @@ import org.eclipse.scanning.event.queues.ServicesHolder;
  * @author Michael Wharmby
  *
  */
-public class GetServerStringResponse implements IQueueReponseStrategy {
+public class GetServerStringResponse extends QueueResponseProcess {
 	
 	private IQueueControllerService queueControl;
 	
-	public GetServerStringResponse() {
+	public GetServerStringResponse(QueueRequest requestBean, IPublisher<QueueRequest> reponseBroadcaster) {
+		super(requestBean, reponseBroadcaster);
 		queueControl = ServicesHolder.getQueueControllerService();
 	}
 
 	@Override
-	public QueueRequest doResponse(QueueRequest request) throws EventException {
+	public QueueRequest process(QueueRequest request) throws EventException {
 		switch (request.getRequestType()) {
 		case COMMAND_SET:		request.setCommandSetName(queueControl.getCommandSetName());
 								break;
