+++ org.eclipse.scanning.event.queues/src/org/eclipse/scanning/event/queues/remote/QueueResponseProcess.java
@@ -1,6 +1,5 @@
 package org.eclipse.scanning.event.queues.remote;
 
-import org.eclipse.scanning.api.event.EventException;
 import org.eclipse.scanning.api.event.core.IPublisher;
 import org.eclipse.scanning.api.event.core.IResponder;
 import org.eclipse.scanning.api.event.core.IResponseProcess;
@@ -14,20 +13,20 @@ import org.eclipse.scanning.api.event.queues.remote.QueueRequestType;
  * stored in the {@link IQueueService}, with access through 
  * {@link IQueueControllerService}. Requests are received as 
  * {@link QueueRequest} beans, with a {@link QueueRequestType}. Based on this 
- * value, an {@link IQueueReponseStrategy} is selected to get the populated 
- * the necessary fields in the request. Finally the {@link QueueRequest} is 
- * passed back to the parent {@link IResponder}. 
+ * value, a class extending {@link QueueResponseProcess} is selected (through 
+ * a strategy pattern) to populate the necessary fields in the request. 
+ * Finally the {@link QueueRequest} is passed back to the parent 
+ * {@link IResponder}. 
  * 
  * @author Michael Wharmby
  *
  */
-public class QueueResponseProcess implements IResponseProcess<QueueRequest> {
+public abstract class QueueResponseProcess implements IResponseProcess<QueueRequest> {
 	
 	private final QueueRequest requestBean;
 	private final IPublisher<QueueRequest> reponseBroadcaster;
-	private IQueueReponseStrategy responder;
 	
-	public QueueResponseProcess(QueueRequest requestBean, IPublisher<QueueRequest> reponseBroadcaster) {
+	protected QueueResponseProcess(QueueRequest requestBean, IPublisher<QueueRequest> reponseBroadcaster) {
 		this.requestBean = requestBean;
 		this.reponseBroadcaster = reponseBroadcaster;
 	}
@@ -41,33 +40,5 @@ public class QueueResponseProcess implements IResponseProcess<QueueRequest> {
 	public IPublisher<QueueRequest> getPublisher() {
 		return reponseBroadcaster;
 	}
-
-	@Override
-	public QueueRequest process(QueueRequest request) throws EventException {
-		
-		
-		//Decide which strategy to use to interrogate the QueueService
-		switch (requestBean.getRequestType()) {
-		case BEAN_STATUS:			
-			responder = new BeanStatusReponse();
-			break;
-		
-		case COMMAND_SET:	// Options COMMAND_SET to JOB_QUEUE_ID all use the GetServerStringResponse
-		case COMMAND_TOPIC:		
-		case HEARTBEAT_TOPIC:	
-		case JOB_QUEUE_ID:			
-			responder = new GetServerStringResponse();
-			break;
-		
-		case SERVICE_START_STOP:	
-			responder = new StartStopResponse();
-			break;
-		
-		default: responder = null;
-		}
-		if (responder == null) throw new EventException("Unsupported QueueRequestType");
-		return responder.doResponse(request);
-	}
-
-
+	
 }
