+++ org.eclipse.scanning.event.queues/src/org/eclipse/scanning/event/queues/remote/package-info.java
@@ -1,8 +1,11 @@
 /**
  * 
- */
-/**
- * @author wnm24546
+ * This package contains classes which allow remote control of the 
+ * {@link IQueueService} through the {@link IQueueControllerService} by a 
+ * client. Most of the interactions are mediated though the 
+ * {@link QueueResponseProcess}. The QueueReponseCreator is the runner of the 
+ * {@link IResponder} associated with the {@link IQueueService} and is 
+ * responsible for creating the correct type of {@link QueueResponseProcess}. 
  *
  */
 package org.eclipse.scanning.event.queues.remote;
